Simplify, simplify, simplify
Our life is frittered away by detail. Simplify, simplify, simplify! I
say, let your affairs be as two or three, and not a hundred or a
thousand; instead of a million count half a dozen, and keep your
accounts on your thumb-nail.
--Henry David Thoureau

Simplifies components and the engine by removing the scene graph. Scene
graphs can probably be implemented as an extension of Component, if you
want them. The unnecessarily complicated things at this stage.

They're probably worth reimplementing in a core library in the future,
since god knows the last thing you want to do when you pick up a new
engine is implement your own scene graph.