Loop through after() calls on Systems backwards
Previously the kernel looped through the before(), update(), and after()
calls on its Systems in the order that the Systems were attached.
However, that makes it difficult to make around-filters, where you can
ensure that something happens truly first and truly last: it could only
happen truly first and then not-truly last, or not-truly first and truly
last.

Looping through the after() calls backwards allows around-filters to
work more intuitively. Incidentally, this is the same decision the Rails
core team arrived on for their implementation of before and after
filters.