More fluid jump controls.
Previously, we only considered the jump key when it was "pressed," as
opposed to "down." This made sense prior to the state machine's
existence, where the keyboard controller needed to only fire these
events carefully. However, now that we have a state machine
controlling exactly when a jump command is allowable, we can attempt to
jump whenever the player has the jump key down -- we'll only jump
when the key is allowed. This makes minor misfires (hit the jump key a
split second too early, while still in the air) work and feel smooth.