Support restricting collisions to physics types
Also fires collision events on both elements that collide, instead of
only alerting the first one to detect the event. Note that this may
result in double collisions: if Alice collides into Bob, and Bob was
also headed on a crash course towards Alice, when Alice's collision is
resolved both she and Bob will be notified and her movement will be
stopped, but Bob's won't be, meaning that Bob will cause another
collision when he gets resolved.

This begins to highlight the need for an overall physical system,
instead of treating every physical object independently.