Update tables to use sparse arrays
Previously, tables always ensured that there were no holes in their
internal index arrays. However, this meant you could have iteration bugs
if one item caused either itself or another item to be removed from the
array mid-iteration, since the length of the array is cached and the
iteration index itself also wouldn't update to reflect that the array
contents had shifted.

This fixes that subtle bug by simply nulling out the attached components
instead of removing them. At the end of each kernel tick, the tables get
compacted and holes are removed.

However, this adds a new wrinkle to getAttached() and getComponents(),
which is that they may have null members. Existing systems need to be
updated to reflect that possibility.