Add smoothing epsilon to Camera class.
Prevents tiny, infinitesimal changes from going on forever. The problem
with the tiny changes is that if you're very, very near a pixel
boundary, at some point you could cross the boundary and snap over, even
when seemingly "at rest" (due to the smoothing over time).

Epsilon values snap the current value to the target value if the
difference is smaller than the epsilon.